Implemented request permission behaviour for Electron and Browser (#231)
* Implemented request permission behaviour for Electron and Browser

Gives a consistent API to call for all three containers, but Electron and
OpenFin automatically grant permission without asking the user.

Defined interface to match the existing standard API in the browser.

* Implement ssf.Notification API with Promise based permission

Also removed the `permission` getter, since it isn't part of the TypeScript
definition on the Window object. For client code it's much cleaner to always
call the promise function rather than check the property to decide whether
to call the promise.