Update demo and change to capture child
Change implementation of capture to capture a screenshot of the visible
child window, instead of the parent window.

Update demo to remove ScreenSnippet and add capture demo to Window.

Add documentation to explain that for the browser, capturing a child
window only works if the document is available from the calling Window,
i.e. they are on the same domain or have permission to access the
domain.