Fixed issue with added frame size in Electron
When electron shows the developer menu, it adds an extra 20/25 pixels to
the min/max height settings. Since the -d option was added to ssf-electron,
the developer menu hasn't been there, so these tests were failing.

For some reason, the OpenFin implementation was modified to also add 20/25
pixels to min/max height, so I've removed that too.

# Conflicts:
#	packages/api-tests/test/window.spec.js